don't use class attributes in matlab wrapper
