+++ gtwrap/matlab_wrapper/utils.py
@@ -1,5 +1,9 @@
 """Mixins for reducing the amount of boilerplate in the main wrapper class."""
 
+import gtwrap.interface_parser as parser
+import gtwrap.template_instantiator as instantiator
+
+
 class CheckMixin:
     """Mixin to provide various checks."""
     # Data types that are primitive types
